#9 refactor
